pyproject.toml
[tool.poetry]
name = "cronjob"
version = "0.0.1"
description = ""
authors = ["chintan sakhiya <chintansakhiya01@gmail.cpm>"]

[tool.poetry.dependencies]
python = "^3.8, <3.11"
hikaru-model-26 = "^1.1.1"


[tool.poetry.dev-dependencies]
robusta-cli = "^0.8.9"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

